Restructure packages/tests to packages/localdev with bun script shortcuts
Rename @march-madness/tests to @march-madness/localdev since the package
is primarily a local dev tool (populate script), not just tests. Move
integration.test.ts from src/ to test/ directory. Add root-level bun
script shortcuts: `bun p:pre`, `bun p:post`, `bun p:grading`.
