+++ packages/localdev/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "types": ["bun-types"]
   },
-  "include": ["src"]
+  "include": ["src", "test"]
 }
