+++ packages/mise.toml
@@ -5,30 +5,30 @@ dir = "{{config_root}}/.."
 run = "bun install"
 
 [tasks.typecheck]
-description = "Typecheck client, web, and tests packages"
+description = "Typecheck client, web, and localdev packages"
 depends = ["install"]
 run = [
   "cd ../packages/client && bun run typecheck",
   "cd ../packages/web && bun run typecheck",
-  "cd ../packages/tests && bun run typecheck",
+  "cd ../packages/localdev && bun run typecheck",
 ]
 
 [tasks."lint::check"]
-description = "Lint check client, web, and tests packages"
+description = "Lint check client, web, and localdev packages"
 depends = ["install"]
 run = [
   "cd ../packages/client && bun run lint:check",
   "cd ../packages/web && bun run lint:check",
-  "cd ../packages/tests && bun run lint:check",
+  "cd ../packages/localdev && bun run lint:check",
 ]
 
 [tasks.build]
-description = "Build client, web, and tests packages"
+description = "Build client, web, and localdev packages"
 depends = ["install"]
 run = [
   "cd ../packages/client && bun run build",
   "cd ../packages/web && bun run build",
-  "cd ../packages/tests && bun run build",
+  "cd ../packages/localdev && bun run build",
 ]
 
 [tasks.test]
